Move files into groovy dir and replace gmaven plugin with gmaven-plus
First I readed that is easier to put all groovy classes into the java 
dir. But with that folder organization, the junit test will not be run.
So I moved all groovy classes into the groovy dir. 

Also I replaced the gmaven plugin with the newer on gmaven-plus plugin.
No I can use the newest version of groovy.